Recommend Remmina for accessing AppVeyor from Linux.
I installed xrdp, and from what I can tell from the man page and the
short screencast (I couldn't find much in the way of documentation for
using the software, only for developing it), the purpose of xrdp is to
make it possible to access the current machine from Remote Desktop. In
other words, it enables you to access your Linux box from your Windows
machine.

In contrast, I was able to use Remmina to access AppVeyor from Ubuntu.

https://remmina.org/