Manually add back figures from pairwise overlaps post.
Because I was gitignoring the supporting CSS/JS files in _posts/,
the figures weren't committed. Thus this only works the first time
when they are moved to public/. In subsequent runs of render_site(),
when they are no longer in _posts/, then they are deleted in public/
and not replaced.

For now, updated the gitignore to allow figures but not CSS/JS. Not
sure how well this will work long-term, especially since other types
of supporting files are possible.